requirements.txt
chardet==4.0.0
langchain==0.0.170
openai==0.27.6
pandas==1.5.3
pypdf==3.8.1
python-dotenv==1.0.0
streamlit==1.22.0
streamlit_chat==0.0.2.2
